Move the prescription details page components to a separate folder
